Move firestore::model::mutation::Overlay into its parents namespace
