+++ Firestore/core/src/local/memory_document_overlay_cache.cc
@@ -29,8 +29,8 @@ namespace local {
 using model::DocumentKey;
 using model::DocumentKeyHash;
 using model::Mutation;
+using model::Overlay;
 using model::ResourcePath;
-using model::mutation::Overlay;
 
 absl::optional<Overlay> MemoryDocumentOverlayCache::GetOverlay(
     const DocumentKey& key) const {
